pc/requirements.txt
Flask==2.0.0
jupyterlab==3.0.12
Keras-Applications==1.0.8
matplotlib==3.4.2
opencv-python==4.5.1.48
pandas==1.2.4
Pillow==8.2.0
seaborn==0.11.1
tensorflow==2.5.0rc2
tqdm==4.60.0